[refactor] redis 패키지 위치/코드 수정
