+++ src/main/java/org/hana/wooahhanaapi/domain/community/service/CommunityService.java
@@ -15,10 +15,10 @@ import org.hana.wooahhanaapi.domain.community.repository.AutoDepositRepository;
 import org.hana.wooahhanaapi.domain.plan.dto.GetMembersResponseDto;
 import org.hana.wooahhanaapi.domain.plan.entity.PlanEntity;
 import org.hana.wooahhanaapi.domain.plan.repository.PlanRepository;
-import org.hana.wooahhanaapi.utils.redis.ValidateAccountPort;
-import org.hana.wooahhanaapi.utils.redis.dto.AccountValidationConfirmDto;
-import org.hana.wooahhanaapi.utils.redis.SaveValidCodePort;
-import org.hana.wooahhanaapi.utils.redis.dto.SendValidationCodeReqDto;
+import org.hana.wooahhanaapi.redis.ValidateAccountPort;
+import org.hana.wooahhanaapi.redis.dto.AccountValidationConfirmDto;
+import org.hana.wooahhanaapi.redis.SaveValidCodePort;
+import org.hana.wooahhanaapi.redis.dto.SendValidationCodeReqDto;
 import org.hana.wooahhanaapi.account.exception.IncorrectValidationCodeException;
 import org.hana.wooahhanaapi.domain.community.dto.*;
 import org.hana.wooahhanaapi.domain.community.entity.CommunityEntity;
