+++ src/main/java/org/hana/wooahhanaapi/redis/SaveValidCodeForMemberPort.java
@@ -1,0 +1,5 @@
+package org.hana.wooahhanaapi.redis;
+
+public interface SaveValidCodeForMemberPort {
+    void saveValidCodeForMember(String phoneNumber, String validCode);
+}
