+++ src/main/java/org/hana/wooahhanaapi/redis/SaveValidCodePort.java
@@ -1,0 +1,5 @@
+package org.hana.wooahhanaapi.redis;
+
+public interface SaveValidCodePort {
+    void saveValidCode(String accountNumber, String validCode);
+}
