+++ src/main/java/org/hana/wooahhanaapi/redis/ValidateAccountPort.java
@@ -1,0 +1,7 @@
+package org.hana.wooahhanaapi.redis;
+
+import org.hana.wooahhanaapi.redis.dto.AccountValidationConfirmDto;
+
+public interface ValidateAccountPort {
+    boolean validateAccount(AccountValidationConfirmDto dto);
+}
