+++ src/main/java/org/hana/wooahhanaapi/redis/ValidateMemberPort.java
@@ -1,0 +1,7 @@
+package org.hana.wooahhanaapi.redis;
+
+import org.hana.wooahhanaapi.redis.dto.MemberValidationConfirmDto;
+
+public interface ValidateMemberPort {
+    boolean validateMember(MemberValidationConfirmDto dto);
+}
