+++ src/main/java/org/hana/wooahhanaapi/redis/dto/AccountValidationConfirmDto.java
@@ -1,0 +1,15 @@
+package org.hana.wooahhanaapi.redis.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class AccountValidationConfirmDto {
+    private String accountNumber;
+    private String validationCode;
+}
