+++ src/main/java/org/hana/wooahhanaapi/redis/dto/SendValidationCodeReqDto.java
@@ -1,4 +1,4 @@
-package org.hana.wooahhanaapi.utils.redis.dto;
+package org.hana.wooahhanaapi.redis.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class AccountValidationConfirmDto {
+public class SendValidationCodeReqDto {
+    private String bankTranId; //하나은행:"001" 1.계좌 생성 시 무조건 "001" 2.회원가입 시 은행 선택
     private String accountNumber;
-    private String validationCode;
 }
