+++ src/main/java/org/hana/wooahhanaapi/sms/port/SMSValidationPort.java
@@ -1,7 +1,6 @@
 package org.hana.wooahhanaapi.sms.port;
 
-import net.nurigo.sdk.message.response.SingleMessageSentResponse;
-import org.hana.wooahhanaapi.utils.redis.dto.MemberValidationConfirmDto;
+import org.hana.wooahhanaapi.redis.dto.MemberValidationConfirmDto;
 
 public interface SMSValidationPort {
     boolean sendOne(String phoneNumber);
