--- src/main/java/org/hana/wooahhanaapi/utils/redis/SaveValidCodePort.java
@@ -1,6 +1,0 @@
-package org.hana.wooahhanaapi.utils.redis;
-
-public interface SaveValidCodePort {
-    void saveValidCode(String accountNumber, String validCode);
-    void saveValidCodeForMember(String phoneNumber, String validCode);
-}
