--- src/main/java/org/hana/wooahhanaapi/utils/redis/ValidateAccountPort.java
@@ -1,7 +1,0 @@
-package org.hana.wooahhanaapi.utils.redis;
-
-import org.hana.wooahhanaapi.utils.redis.dto.AccountValidationConfirmDto;
-
-public interface ValidateAccountPort {
-    boolean validateAccount(AccountValidationConfirmDto dto);
-}
