--- src/main/java/org/hana/wooahhanaapi/utils/redis/ValidateMemberPort.java
@@ -1,7 +1,0 @@
-package org.hana.wooahhanaapi.utils.redis;
-
-import org.hana.wooahhanaapi.utils.redis.dto.MemberValidationConfirmDto;
-
-public interface ValidateMemberPort {
-    boolean validateMember(MemberValidationConfirmDto dto);
-}
