+++ src/test/java/org/hana/wooahhanaapi/service/CommunityServiceTest.java
@@ -15,8 +15,8 @@ import org.hana.wooahhanaapi.domain.member.entity.MemberEntity;
 import org.hana.wooahhanaapi.domain.member.exception.UserNotFoundException;
 import org.hana.wooahhanaapi.domain.member.repository.MemberRepository;
 import org.hana.wooahhanaapi.domain.member.service.MemberService;
-import org.hana.wooahhanaapi.utils.redis.SaveValidCodePort;
-import org.hana.wooahhanaapi.utils.redis.ValidateAccountPort;
+import org.hana.wooahhanaapi.redis.SaveValidCodePort;
+import org.hana.wooahhanaapi.redis.ValidateAccountPort;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
